==> 0. Numpy.ipynb <==
# 0. Numpy
NumPy is Python's library that facilitates processing matrices or generally large multidimensional arrays. In addition to the data structure, NumPy provides efficiently implemented features for numerical computation.

import numpy as np

arr1 = np.array([1,2,3,5,7,9,11]) # numpy 배열 생성
arr1

type(arr1) # ndarray : N-dimensional Array

arr1.shape # .shape : 데이터의 차원 출력 (1차원)

arr1.size # .size : 데이터의 갯수 출력

arr2 = np.array([[1,2,3,4],[5,6,7,8]])
arr2

arr2.shape # 2차원

arr2.size

### Numpy Method
> np.full(n, a) : n개의 데이터를 a로 채움 

> np.zeros(n) : n개의 데이터를 0으로 채움

> np.ones(n) : n개의 데이터를 1로 채움

> np.random.randint(a, b) : a~b 사이의 정수를 받아옴

> np.random.random(n) : 0~1사이의 실수 n개를 받아옴

> np.arange(n, m, s) : n부터 m-1까지 간격이 s인 값들이 담긴 배열 생성

arr1 = np.full(6,2) # 6개의 데이터를 2로 채움
arr1

arr2 = np.zeros([2, 6]) # 2행 6열의 데이터를 0으로 채움
arr2

arr2 = np.zeros([2, 6], dtype = int) # 데이터 타입 설정 가능
arr2

arr3 = np.ones(5)
arr3

arr4 = np.random.randint(1, 100)
arr4

arr5 = np.random.random(5)
arr5

arr6 = np.arange(10)
arr6

arr1 = np.arange(5)
arr2 = np.arange(5, 10)

## Numpy 연산
기본적으로 numpy배열의 연산은 각 행과 열에 대응되는 숫자끼리 연산을 수행한다.

arr1

arr2

arr1 + arr2

arr2 - arr1

arr1 * arr2

arr1 / arr2

arr2 > arr1

## Numpy 통계값

arr1 = np.array([13,2,6,5,12,3,5,1,5])
arr1

arr1.max() # 최댓값

arr1.min() # 최솟값

arr1.mean() # 평균값

arr1.std() # 표준편차

arr1.var() # 분산

np.median(arr1) # 중앙값
