==> 1. Series.ipynb <==
# 1. Series
One-dimensional ndarray with axis labels (including time series).

Labels need not be unique but must be a hashable type. The object supports both integer- and label-based indexing and provides a host of methods for performing operations involving the index. Statistical methods from ndarray have been overridden to automatically exclude missing data (currently represented as NaN).

https://pandas.pydata.org/docs/reference/api/pandas.Series.html

import pandas as pd

## Constructing Series from a dictionary with an Index specified

month = pd.Series([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12])
month # Series Object

month[0] # Jan

## Series 객체에 Index 지정

month = pd.Series([1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12], index = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'])
month

month['Jan']

month['Mar']
