==> 2. DataFrame.ipynb <==
# 2. DataFrame
> class pandas.DataFrame(data=None, index=None, columns=None, dtype=None, copy=None

Two-dimensional, size-mutable, potentially heterogeneous tabular data.

Data structure also contains labeled axes (rows and columns). Arithmetic operations align on both row and column labels. Can be thought of as a dict-like container for Series objects. The primary pandas data structure.

https://pandas.pydata.org/docs/reference/api/pandas.DataFrame.html

## Prepare Data

data = {
    'Name' : ['Tom','Noah','Lucas','Mia','Amelia','Sophia', 'James', 'Ava'],
    'Height' : [175, 190, 185, 165, 170, 160, 183, 172],
    'Weight' : [80, 90, 83, 50, 56, 60, 72, 55],
    'School' : ['MIT', 'MIT', 'Harvard', 'MIT', 'Harvard', 'MIT', 'Harvard', 'MIT'],
    'Gender' : ['M', 'M', 'M', 'F', 'F', 'F', 'M', 'F'],
    'Python' : [80, 90, 30, 66, 72, 93, 100, 76],
    'Java' : [75, 92, 25, 54, 77, 95, 97, 73],
    'Javascript' : [77, 95, 15, 59, 62, 90, 100, 70],
    'C#' : [97, 87, 23, 61, 74, 88, 95, 66],
    'C++' : [92, 80, 13, 54, 72, 85, 91, 68],
    'HTML' :[72, 93, 50, 54, 69, 82, 99, 73],
}
data

data['Name']

data['School']

## Constructing DataFrame from a dictionary.

import pandas as pd
df = pd.DataFrame(data)
df

df.dtypes # 데이터 타입 확인

## DataFrame 객체에 Index 지정

df = pd.DataFrame(data, index = ['No.1','No.2','No.3','No.4','No.5','No.6','No.7','No.8'])
df

df.index

df.columns

## 특정 데이터에 접근

df['Name']

df[['Name','School']]

# 원하는 Columns만을 가지고 DataFrame 객체를 생성할 수 있다.
tmp = pd.DataFrame(data, columns=['Name','Gender','School'])
tmp

# DataFrame 파일 저장 및 불러오기

## .csv 파일로 저장

df.to_csv('info.csv', index=False) # 앞서 설정한 Index 빼고 저장
# df.to_csv('info.csv',encoding='utf-8-sig') # 데이터에 한글이 있다면, encoding을 해주어야 한다.

## .txt 파일로 저장

df.to_csv('info.txt', sep='\t', index=False) # 분리자 (tap으로 구분 된 텍스트 파일)

## .xlsx 파일로 저장

df.to_excel('info.xlsx', index=False)

***
## .csv 파일 열기

df = pd.read_csv('info.csv')
df

df = pd.read_csv('info.csv', skiprows=3) # 3개의 row 스킵하고 불러옴
df

df = pd.read_csv('info.csv', skiprows=[1,2,6]) # 지정한 row 스킵하고 불러옴
df

df = pd.read_csv('info.csv', skiprows=3, nrows=4) # 처음 3개 row스킵 후 4개의 row불러옴
df

## .txt 파일 열기

df = pd.read_csv('info.txt', sep='\t') # 분리자로 구분하여 저장했기에 sep옵션을 추가하여 불러옴.
df

## .xlsx 파일 열기

- 해당 기능은 잠재적인 안정성 이슈때문에 지원을 중단했으며, 사용하고 싶으면 openpyxl 라이브러리를 사용해야 한다.

# df = pd.read_excel('info.xlsx')
