==> 3. DataFrame_Index.ipynb <==
# 3. DataFrame Index
데이터에 접근할 수 있는 주소값인 인덱스에 대한 함수들
> df.index : 인덱스의 정보를 보여줌

> df.index.name : 인덱스 이름 설정 가능

> df.reset_index() : 앞서 설정한 인덱스 초기화

> df.set_index() : 인덱스 설정

> df.sort_index() : 인덱스를 기준으로 오름차순/내림차순 정렬


import pandas as pd

df = pd.read_csv('info.csv')
df

## 인덱스 설정

df.index = ['No.1','No.2','No.3','No.4','No.5','No.6','No.7','No.8']
df

df.index # 인덱스가 무엇이 있는지 보여줌

df.index.name = 'Number' # 인덱스 이름 설정
df

df.reset_index(inplace=True) # 인덱스 초기화 (inplace=True)로 해주어야 원래 DataFrame객체에 적용
df

df.set_index('Number', inplace=True) # 지정한 Column으로 인덱스 지정
df

df.reset_index(drop=True, inplace=False) # (drop=True : 현재 설정된 인덱스 삭제)

df.sort_index() # 인덱스를 기준으로 오름차순 정렬

df.sort_index(ascending=False) # 인덱스를 기준으로 내림차순 정렬

# 추가한 인덱스를 바탕으로 파일 저장
df.to_csv('info.csv')

df = pd.read_csv('info.csv')
df
