==> 4. Data Analysis.ipynb <==
# 4. Data Analysis

import pandas as pd
df = pd.read_csv('info.csv')
df

df = pd.read_csv('info.csv', index_col='Number') # Number을 인덱스로 설정하여 불러옴
df

type(df) # DataFrame Object 확인

df.info() # 데이터 정보를 확인

df.index # index 확인

df.columns # columns 확인

df.dtypes # 데이터 타입 확인

df.shape # (row, column) 확인

df.values # 데이터 값 확인

df.describe() # Hashable 데이터의 통계치 확인

df.head() # 첫 5개의 데이터 확인

df.tail() # 끝 5개의 데이터 확인

## 특정 column(Series)의 통계치 확인

df['Python'].describe() # 통계치 확인

df['Python'].max() # 최댓값

df['Python'].min() # 최솟값

df['Python'].nlargest(3) # 가장 Python 점수가 높은 3명 확인

df['Python'].mean() # 평균

df['Python'].count() # NaN이 아닌 데이터 확인

df['School'].unique() # unique한 데이터 확인
