==> 5. Data Selection.ipynb <==
import numpy as np
import pandas as pd

df = pd.read_csv('info.csv',index_col=0)
df

df['Name']

df['Gender']

df[['Height','Weight']]

df.columns

## index로 특정 Column에 접근

df.columns[0]

df.columns[1]

df[df.columns[0]] # == df['Name']

df[df.columns[-1]] # 맨 끝의 값을 가져옴

## Slicing

df['Name'][0:4]

df[['Name','Gender']][0:4]

df[3:]

***
# pandas.DataFrame.loc
#### property DataFrame.loc
Access a group of rows and columns by label(s) or a boolean array.

> df.loc[row_sel, col_sel]

이름을 이용하여 원하는 row 에서 원하는 col 선택

df.loc['No.1'] # row_sel

df.loc['No.5','Python'] # row_sel, col_sel

df.loc[['No.1','No.2'], 'School'] # row_sel, col_sel

df.loc[['No.1', 'No.6'], ['School','Gender']] # row_sel, col_sel

df.loc['No.2':'No4', 'Python':'HTML'] # row_sel, col_sel

***
# pandas.DataFrame.iloc
#### property DataFrame.iloc
Purely integer-location based indexing for selection by position.

> df.iloc[row_sel, col_sel]

위치를 이용하여 원하는 row 에서 원하는 col 선택

df.iloc[0] # row_sel

df.iloc[4] # row_sel

df.iloc[1:3] # row_sel

df.iloc[2, 3] # row_sel, col_sel

df.iloc[[2, 3], 4] # row_sel, col_sel

df.iloc[[2, 3], [4, 5]] # row_sel, col_sel

df.iloc[:3, 1:5] # row_sel, col_sel

## 조건에 해당하는 값 선택

df['Height'] > 180

filt = df['Height'] > 180 
df[filt] # No.2, No.3, No.7 (True) 값 만을 반환

df[~filt] # 2, 3, 7 이외의 값 반환

df.loc[filt, 'Python'] # row_sel, col_sel

df.loc[filt, ['School','Gender','Python']] # row_sel, col_sel

filt1 = df['Height'] > 180
filt2 = df['School'] == 'MIT'

### 조건 : &

df.loc[filt1 & filt2] # 특정 조건(filtering)을 만족하는 데이터 찾기

### 조건 : |

df.loc[filt1 | filt2]

filt = df['School'].str.startswith('M') # M으로 시작하는 학교를 다니는 사람에 대한 데이터
df[filt]

df[~filt] # M으로 시작하는 학교를 다니는 사람 제외한 데이터

sch = ['MIT','Harvard']
filt = df['School'].isin(sch) # 학교가 MIT 또는 Harvard인 사람에 대한 데이터
df[filt]

df['Name'].str.lower() # 모두 소문자로 표현

df['Name'].str.upper() # 모두 대문자로 표현

df['Name'].str.contains('me', na=False) # 이름에 me가 포함되는 사람에 대한 데이터 (NaN에 대해서는 False로 설정)

filt = df['Name'].str.contains('me', na=False)
df[filt]
